[AIRFLOW-3458] Move connection tests
